Project/requirements.txt
streamlit==1.42.2
google-cloud-firestore==2.20.2
supabase==2.15.1